fix process output forwarding

+++ src/main/java/com/ampznetwork/discordbot/BotCommands.java
@@ -5,17 +5,21 @@ import com.ampznetwork.discordbot.io.DiscordChannelOutputStream;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.Value;
 import lombok.experimental.FieldDefaults;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
 import org.comroid.api.attr.Named;
 import org.comroid.api.func.util.Command;
-import org.comroid.api.func.util.DelegateStream;
+import org.comroid.api.os.OS;
 import org.comroid.api.text.StringMode;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.Executors;
 
@@ -52,6 +56,19 @@ public class BotCommands {
         }
     }
 
+    @Value
+    private static class Forwarder implements Runnable {
+        InputStream  input;
+        OutputStream output;
+
+        @Override
+        @SneakyThrows
+        @SuppressWarnings("InfiniteLoopStatement")
+        public void run() {
+            while (true) input.transferTo(output);
+        }
+    }
+
     @Command(permission = "8")
     public static class execute {
         private static final String PROCESS_FINISHED = "Process finished with exit code ";
@@ -59,30 +76,31 @@ public class BotCommands {
         @Command(permission = "8")
         public static String shell(MessageChannelUnion channel, @Command.Arg(stringMode = StringMode.GREEDY) String script) {
             try {
-                var tmp = File.createTempFile("script", ".sh");
-                try (var fis = new FileOutputStream(tmp)) {
-                    fis.write(script.getBytes(StandardCharsets.US_ASCII));
-                }
+                File           tmp = null;
+                ProcessBuilder proc;
+                if (!OS.isWindows) {
+                    tmp = File.createTempFile("script", ".sh");
+                    try (var fis = new FileOutputStream(tmp)) {
+                        fis.write(script.getBytes(StandardCharsets.US_ASCII));
+                    }
+                    proc = new ProcessBuilder("bash", tmp.getAbsolutePath());
+                } else proc = new ProcessBuilder("cmd.exe", "/c", script);
 
-                try {
-                    var proc = new ProcessBuilder("bash", tmp.getAbsolutePath());
-                    var pool = Executors.newCachedThreadPool();
-
+                try (var pool = Executors.newCachedThreadPool()) {
                     try (
-                            var dcIn = new DiscordChannelInputStream(channel);
-                            var dcOut = new DiscordChannelOutputStream(channel, "ℹ️ - ");
+                            var dcIn = new DiscordChannelInputStream(channel); var dcOut = new DiscordChannelOutputStream(channel, "ℹ️ - ");
                             var dcErr = new DiscordChannelOutputStream(channel, "⚠️ - ");
                     ) {
                         var exec = proc.start();
-                        DelegateStream.redirect(dcIn, exec.getOutputStream(), pool);
-                        DelegateStream.redirect(exec.getInputStream(), dcOut, pool);
-                        DelegateStream.redirect(exec.getErrorStream(), dcErr, pool);
+                        pool.execute(new Forwarder(dcIn, exec.getOutputStream()));
+                        pool.execute(new Forwarder(exec.getInputStream(), dcOut));
+                        pool.execute(new Forwarder(exec.getErrorStream(), dcErr));
 
                         return PROCESS_FINISHED + exec.waitFor();
                     }
                 } finally {
-                    //noinspection ResultOfMethodCallIgnored
-                    tmp.delete();
+                    if (tmp != null)//noinspection ResultOfMethodCallIgnored
+                        tmp.delete();
                 }
             } catch (InterruptedException | IOException e) {
                 throw new RuntimeException("A fatal unexpected error has occurred", e);
